pyproject.toml
[project]
name = "geoguessr-ai"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
license-files = ["LICENSE"]
dependencies = [
    "fiona>=1.10.1",
    "geopandas>=1.1.1",
    "ipython>=9.5.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "pydeck>=0.9.1",
    "shapely>=2.1.2",
    "tqdm>=4.67.1",
    "boto3>=1.40.36",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "pillow>=11.3.0",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "uvicorn[standard]>=0.37.0",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "pandas>=2.3.3",
    "datasets>=4.2.0",
    "torch>=2.8.0",
    "transformers>=4.57.1",
    "torchvision>=0.23.0",
    "accelerate>=1.10.1",
    "rasterio>=1.4.3",
    "timm>=1.0.21",
    "torchsummary>=1.5.1",
    "ipykernel>=7.1.0",
    "s3fs>=2025.9.0",
    "mmpretrain>=1.2.0",
    "botocore>=1.40.61",
    "wandb>=0.22.3",
    "tensorboard>=2.20.0",
]


[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
]

training = [
    "transformers>=4.57.1",
    "torchvision>=0.23.0",
]

docs = [
    "mkdocs>=1.6.1",
    "mkdocs-api-autonav>=0.3.0",
    "mkdocs-autoapi>=0.4.1",
    "mkdocs-awesome-nav>=3.1.2",
    "mkdocs-material>=9.6.16",
    "mkdocstrings>=0.30.0",
    "mkdocstrings-python>=1.16.12",
]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "uv"
update_changelog_on_bump = true
major_version_zero = true
